chore(payments-server): move AlertBar component files into directoy
- refs #3218
