+++ packages/fxa-payments-server/src/components/AlertBar/index.scss
@@ -17,7 +17,7 @@
 }
 
 .alert span.checked::before {
-  background-image: url('../images/check.svg');
+  background-image: url('../../images/check.svg');
   background-position: 0 3px;
   background-repeat: no-repeat;
   content: '\00a0';
