+++ packages/fxa-payments-server/src/components/AlertBar/index.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { AlertBar } from './AlertBar';
+import { AlertBar } from './index';
 
 storiesOf('components/AlertBar', module).add('basic', () => (
   <AlertBar className="alert">This is an alert.</AlertBar>
