+++ packages/fxa-payments-server/src/components/AlertBar/index.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 
-import { AlertBar } from './AlertBar';
+import { AlertBar } from './index';
 
 afterEach(cleanup);
